use pdr-utils and refactor
