+++ requirements.txt
@@ -1,5 +1,1 @@
-web3
-eth-account
-pathlib
-eth-keys
-pdr-utils
+pdr-utils==0.0.12
